Dermalife/pyproject.toml
[tool.poetry]
name = "dermalife"
version = "0.1.0"
description = "This package allow to get disccount in automatic"
authors = ["Juan David Duque <jduque0289@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
requests = "^2.32.3"
woocommerce = "^3.0.0"
python-dotenv = "^1.1.0"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
